// src/common/im/enums.ts
export enum Command {
  COMMAND_UNKNOW = 0,
  COMMAND_HANDSHAKE_REQ = 1,
  COMMAND_HANDSHAKE_RESP = 2,
  COMMAND_AUTH_REQ = 3,
  COMMAND_AUTH_RESP = 4,
  COMMAND_JOIN_GROUP_REQ = 5,
  COMMAND_JOIN_GROUP_RESP = 6,
  COMMAND_JOIN_GROUP_NOTIFY_RESP = 7,
  COMMAND_EXIT_GROUP_NOTIFY_RESP = 8,
  COMMAND_CHAT_REQ = 9,
  COMMAND_CHAT_RESP = 10,
  COMMAND_START_SHOW_REQ = 11,
  COMMAND_START_SHOW_RESP = 12,
  COMMAND_END_SHOW_REQ = 13,
  COMMAND_END_SHOW_NOTIFY_RESP = 14,
  COMMAND_HEARTBEAT_REQ = 15,
  COMMAND_HEARTBEAT_RESP = 16,
  COMMAND_CLOSE_REQ = 17,
  COMMAND_CLIENT_PAGE_REQ = 18,
  COMMAND_CLIENT_PAGE_RESP = 19,
  COMMAND_LOGIN_REQ = 20,
  COMMAND_LOGIN_RESP = 21,
  COMMAND_CANCEL_MSG_REQ = 22,
  COMMAND_CANCEL_MSG_RESP = 23,
  COMMAND_CALL_REQ = 24,
  COMMAND_CALL_RESP = 25,
  COMMAND_REMIND_PUSH = 26,
  COMMAND_ONLINE_PUSH = 28,
  COMMAND_OFFLINE_PUSH = 29,
  COMMAND_CLEAR_REMIND_REQ = 27,
  COMMAND_ONLINE_NOTIFY_SUBSCRIBE_REQ = 30,
}

export enum Code {
  OK = 0,
  FAIL = -1,
}

export enum Role {
  CUSTOMER = 0,
  WAITER = 1,
}